Moved texture scaling to separate file
